Initial invocation image build document
This document includes details on how the invocation image is constructed.
It includes details on how the porter.yaml is used to determine which dependnecies and
mixins to include in the bundle, how the mixins are used to contribute to the resulting
Dockerfile and how dependencies are included in a bundle. Throughout the document, it
uses several example porter.yaml files and the resulting `porter build` command output
to explain each of these scenarios.

Closes #1