Enhances how we're handling slice processing for param/credential/output sourcing
This PR updates how we handle slice elements for doing parameter/credential/ouput sourcing.
We probably should think about revisitng the entire process for this in the future.

In summary, we are using reflectwalk to handle walking the manifest. This involves handling a couple
call backs, notably "SliceElement" and "MapElement". The Slice Element assumed that it would see something
of the form source: <x.y.z> as a string. In reality, if that's not wrapped in quote, Go and YAML see that
as a Map, but the map callback doesn't get called to handle that. This was resulting in the block not getting
rewritten and the exec mixin was exploding (probably any would have). This PR updates the logic of the slice elem
processing to check and see if it's a map and handle it in a similar manner to a top level map entry in the yaml.

Closes: #158