Inject porter-debug param into all bundles
This will let us handoff a --debug flag from porter install into the
bundle without the user having to manually specify the parameter.

So the user can run `porter install --debug` and porter will handle
automatically turning on the porter-debug parameter when it executes the
bundle, and then `porter run` will pick it up and bind its `--debug`
flag to that automatically.

This is just the first step of defining the parameter for all porter
built bundles.