Put porter interfaces in named fields
Previously these were in embedded fields which caused collisions, as
both mixins and cnab actions had "Install" or "Uninstall". Also it was
confusing when porter needed to also have a composite function that
handled installing a bundle that wrapped up functionality from the CNAB
InstallBundle function. It was hard to know what to call when.

So I've moved the interfaces for the CNAB and Mixin providers into
named fields so that it's more clear which are the methods owned by the
Porter struct (and should be used by the cli).