install mixin from url (#265)
* Wire the porter mixin install command

* Put porter interfaces in named fields

Previously these were in embedded fields which caused collisions, as
both mixins and cnab actions had "Install" or "Uninstall". Also it was
confusing when porter needed to also have a composite function that
handled installing a bundle that wrapped up functionality from the CNAB
InstallBundle function. It was hard to know what to call when.

So I've moved the interfaces for the CNAB and Mixin providers into
named fields so that it's more clear which are the methods owned by the
Porter struct (and should be used by the cli).

* Implement installing a mixin from a URL

* Document how to publish a mixin

Explain how to publish a mixin that works with porter install

* Improve doc page formatting

* Close gap between page title and initial content
* Title should be h1

* Add unit tests for installing a mixin

* Incorporate review feedback

* --url is required

* Cleanup mixin if we can't install it properly

* Make a test that fails on linux, darnit

* Comment out test since it requires non-root