Do not encode -runtime into mixin filename
Instead of requiring the runtime binary for the mixin
to be uploaded as foo-runtime-linux-amd64, allow the client mixin
for linux, which is foo-linux-amd64, to be reused as there is
no difference.

This sets us up anyway for when porter needs to support a customizable
runtime image that could be a different os/arch.