Add dockerfile template to manifest
When a dockerfile template is specified in the manifest, instead of
using the default Dockerfile template (which uses debian:stretch for the
FROM), the user can supply their own template Dockerfile that they
commit alongside the porter.yaml file. During porter build, we use that
as the template, and appending lines to it.

It is the responsibility of the user to ensure that it is a suitable
image. For example, FROM scratch wouldn't work well because it doesn't
provide root ssl certificates, but FROM ubuntu would be just fine.