Add Credentials Skeleton and Generate Flow (#328)
* Add Credentials Skeleton and Generate Flow

This PR adds skeleton stuff to the porter cmd/ stuff to add credentials commands.
This PR only wires up generate, however, since it's already pretty large. Will handle
the others in follow-ons. The generate command is similar to Duffle in that it uses
survey, but the text and prompts are different.