Fix canary publish (#340)
* Add canary entry to mixin feed

An entry for the canary version (the permalink not the funky version
number) needed to be in the mixin feed but wasn't being added because it
wasn't in a directory in bin/mixins/exec/canary before we ran porter
mixin feed generate.

I also noticed that we lost uploading the mixin binaries to the latest
directory in azure storage. That is a nice thing to have in case someone
needed to download them directly using the permalink. So that is fixed
too.

* Add canary tests to the mixin feed test suite

I've added tests to ensure that having canary (a non-semver) in the
mixin feed didn't cause any regressions.