Stamp bundle.json with a digest of porter.yaml
We are only stamping with the on disk representation of the porter.yaml,
not the dirty in-memory representation. We don't care about in-mem, this
digest will be used later to determine if a rebuild is needed because
the porter.yaml on disk is out of sync with the bundle.json.