Improve unit test setup of porter home
Don't copy the binaries from the real bin on the local file system.
Instead just touch the files that we expect to exist, since they aren't
going to be executed anyway.

This avoid problems where a dirty dev bin/ directory causes unit tests
to fail because of unexpected directories and files existing when they
shouldn't.