Hide the local cnab directory
This hides the local cnab directory that porter generates during the
build phase from the user. The goal is to avoid giving the user
expectations about the directory and its contents.

In addition, we now delete and recreate the generated files, .cnab and
Dockerfile, before we do a build to avoid any odd bugs.

Porter also now provides a build argument, BUNDLE_DIR, to the Dockerfile
template during porter build, so that the user doesn't need to be as
aware of /cnab/app. Our template now uses $BUNDLE_DIR to encourage using
the argument instead of the hardcoded value.

I've also tweaked our gitignore to take advantage of the new naming
rules.