Use operation's options struct instead of print opts
In some places we were directly using the printer options instead of the
parent operations struct when calling an operation. Eventually the list
operation is going to have more parameters, so I wanted to switch to
using the operation's struct now and follow the existing pattern of
passing the entire operation's struct around instead of individual
parameters to operations.