Split manifest in two: porter.yaml and the runtime
Manifest: The struct representing porter.yaml along with the logic to
validate it.

RuntimeManifest: The logic and extra state for keeping the running
represetantion of the manifest as we execute the bundle.

This is just the first step to see how this works for us without causing
extra work or refactoring for the sake of refactoring.