Move outputs pkg into config
I was getting a lot of cyclical import problems because I kept wanting
to use the very useful outputs package from other packages. Since it is
operating on files in the config directory, it seems like a natural fit
to move it into the config package. Hopefully the applies to logic will
move into cnab-go soon anyway.