Add troubleshooting hook to porter schema
When porter.json is present in PORTER_HOME, the porter schema command
will return it, instead of generating a schema. This is intended to
assist with troubleshooting VS Code. It will allow someone to tweak the
schema without modifying porter or the mixins and quickly iterate inside
of VS Code, and then we can update the Go code to match the desired
schema later.