Rename mixins delete to mixins uninstall
For consistency we want to use commands that pair together. add/delete pair, but we are using
install/delete which (just see helm2) causes people to guess the wrong verb. It's
also inconsistent with the verbs we use elsewhere in porter which is already
install/uninstall.

Sorry, that is my fault when I created the issue for the enhancement!