Update Porter to handle relocation mapping on pull/operations (#719)
* Update Porter to handle relocation mapping on pull/operations

This PR updates Porter to handle a relocation mapping if it is generated on a bundle pull. If generated, it is stored in the bundle cache and then passed along via an oper config to add the file to an operation before it is run. Also overwrites the invocation image, if it is found in the relocation mapping