Refactor claims plugin to be a storage plugin
Instead of having a credential plugin, and a claims plugin, etc this
refactors the plugin POC so that there is a single storage plugin with
the intent of being able to offload the entire porter home directory
contents (except the config file) to the plugins.

For example everything that we currently store today on the filesystem
could be stored in blob storage or a database.

This replies upon a unmerged branch with changes to the crud.Store
interface and other changes to support this.

I have also stubbed in a new plugin for secrets, though it isn't
completely plumbed through and will be completed in a separate PR. At
the moment all secrets, for example the values that we resolve
credentials from but could eventually be parameter values too among
other things, are coming from a host plugin embedded in porter.