Switch from -ing to simple verbs in docs
Simplify docs and use consistent verbs in the titles by prefering
imperative verbs over gerunds. For example, Distribute Bundles instead
of Distributing Bundles.

Shorter text and easier for people to read, especially when English
isn't their first language. It also makes it easier for us to be
consistent with our docs, we can pick the simple form of the verb and
just go with that everywhere.