Preserve name casing in credentials show
porter credentials show HELLO should query the backend credentials store
for a credential set named "HELLO" and not force it to lower-case. The
CNAB spec allows for credential sets that use upper-case, and when you
generate credential sets for bundles with upper-case letters, it
defaults to the name of the bundle. So we end up with credential sets
that depending on the backing store, cannot be read due to forcing to
lower-case.

I checked and the other only place we do this is for mixins. I'm going
to leave that as-is for now, since perhaps limiting mixins to all lower
case is useful? I'll make a follow-up issue to consider if we need to
change that too.