Update min version tested by the install script check
Now that we are using porter plugins install in our install script, I
had to bumpb the min version of porter that we use to test the install
script in our CI pipeline so that it's using a version of porter that
_has_ that command.

I've updated it to use the next version number that we are going to tag
in a few minutes. So this may fail the first time it runs but will pass
on subsequent runs after the tag is published.