Install azure by default (#901)
* Fix missing newline when installing packages

We lost the \n when printing the success message after installing a
mixin or plugin. It wasn't obvious until I went to update the install
script and saw the messages printing all on the same line.

* Update install page

* Include information about plugins
* Do not list each mixin/plugin included in the install script, link to
  the index page instead.

* Install the azure plugin by default

* Fix typos

* Install arm mixin by default

Install the arm mixin by default instead of the azure now that
the azure mixin has been deprecated.

* Update min version tested by the install script check

Now that we are using porter plugins install in our install script, I
had to bumpb the min version of porter that we use to test the install
script in our CI pipeline so that it's using a version of porter that
_has_ that command.

I've updated it to use the next version number that we are going to tag
in a few minutes. So this may fail the first time it runs but will pass
on subsequent runs after the tag is published.