Explain why we need credential sets
Add explanation for why we can define credential sources in porter.yaml
from Slack to a new Q & A section on the Credentials page.

Rename the page to just "Credentials" to match how other pages are named
and update references to the page, adding a redirect for the old name.

Fix whitespace on inline <code> blocks so they don't wrap