Explain allow-docker-host-access setting
Update the configuration page with how to use the
allow-docker-host-access setting as either an environment variable or
configuration setting. It is not yet available as a command flag.

Also note that based on how the Azure Cloud Shell driver is implemented
today (using ACI as the host environment for running the porter commands
such as porter install), this setting won't work yet in that
environment.