Trickle down changes to Cache function signature
I've had a change of heart about returning *struct, error from functions.
The cache returns a struct, error instead, and this commit is a
follow-up to the previous to switch pointers/struct returns to handle
the change. In another follow-up later we can slowly change the
codebase to follow this pattern everywhere.

The idea being a pointer should be used when the consumer needs a
mutable reference (like we do with context) or when they _may_ get a nil
reference which should be checked. If they are guaranteed to get a value
back, then the signature should convey that.

I split up the commits into the big change, and then a ripple of message
with returning empty structs instead of nil, or passing pointers vs
structs, etc for easier reviewing.