Consolidate test runtime / cnabprovider structs
There were two test implementations for the cnab runtime. I have
consolidate the two and made it easier to initialize them for unit and
integration testing.

By default unit tests use the debug driver now, instead of completely
bypassing the runtime entirely, which helps us exercise more of our cnab
runtime code. For example a claim will now be created at the end of
install, where before it wasn't.

This did affect the tests that expected stub data, and some needed to be
updated to actually be based on a real test file.

We also don't need to manually swap out a mock cache anymore, and can
use the built-in cache in the TestPorter instance and only Mock error
cases as needed.