Fix bundle load logic in dependencies code
Previously we were accidentally ignoring errors when bundle load failed
in the dependencies code. This helped cover up not handling when the
bundle was supplied through just an instance name. In the latest
release, I fixed the lack of error handling but didn't catch that the
code wasn't loading bundles from the claim name too (just from the file
or tag). This adds in logic to also load the bundle from an instance
name.