Migrate claim data on read
The claim spec made a breaking change, renaming claim.Name to
claim.Installation. This injects a migration into our claim storage
layer so that when we read a claim, we detect older claims and
automatically migrate them.

The old Name field is preserved so that the claim is backwards
compatible with older CNAB clients.