Selectively replace the org when moving images (#1047)
* Selectively replace the org when moving images

Don't use a string replace because when the same text is used in the org
and the image name and the registry, it can end up changing more text
than it should.

Don't replace the org when there isn't one present. Some registries
embed the org in the domain, e.g. getporter.azurecr.io, in which case
the path of the tag is the name of the bundle, and we shouldn't be
appending that to the images.

* Doc artifacts publish --archive creates

* Document the intermediate artifacts created by porter publish --archive
creates.
* Clarify the documentation around how images are named when published.

* Incorporate feedback

* Apply suggestions from code review

Co-authored-by: Vaughn Dice <vadice@microsoft.com>

Co-authored-by: Vaughn Dice <vadice@microsoft.com>