Change defaulting of bundle/image tags
When bundle tag is omitted, default to the bundle version with a 'v'
prefixed. For example:

version: 1.2.3
tag: example.com/mybun

This bundle publishes to example.com/mybun:v1.2.3

When the invocation image is omitted, default to the same tag as the
bundle. In the example above, the invocation image pushed would be
example.com/mybun-installer:v1.2.3. This addresses the confusion where
the invocation image would not have a 'v' prefix but the bundle would
because the invocation image's tag was based on the bundle version.