Add contributors file
This works around a bug/limitation in GitHub where you lose your
contributor status to the project when you disassociate the email that
you contributed with from your GitHub profile. Basically when you change
jobs, it looks like you never contributed which is unacceptable to us.

So we will ensure contributors are acknowleged and thanked via a
Contributors file and properly link it so people know to use it.