Use latest claims code in cnab-go (#1065)
* Use latest claims code in cnab-go

This gets porter using the latest claims spec that is now supported in
cnab-go. It doesn't take advantage yet of new features that we could do
with the additional data, just ensures that we still build and work.

* Apply suggestions from code review

Co-authored-by: Vaughn Dice <vadice@microsoft.com>

* Fix test-cli

* Set PORTER_HOME in integration tests on plugins

The integration tests don't use the PORTER_HOME env var, they
programmatically set the porter home directory on the porter app. So
when Porter creates a plugin, we need to explicitly set the home dir on
the plugin when it's instantiated since it won't get PORTER_HOME
automatically.

* Fix show instance status

Bump cnab-go to get fix for displaying instance status. We need the
entire installation history so that we have the installation time (the
first claim), and ReadInstallation needs to read claims AND results.

* Update slide reference to claim file

* Resolve merge conflicts with paramset feature

* Incorporate review feedback

* Change error handling when we can't find the install time
* Move definition of common runtime functions
* Remove accidental early return

* Consolidate logic for cnab provider

ONE FUNCTION TO RULE THEM ALL!

* Move UseFilesystem onto TestContext

* bump cnab-go

* Final cnab-go bump v0.13.0-beta1

Co-authored-by: Vaughn Dice <vadice@microsoft.com>