Claim data migration (#1090)
* Migrate to new claim layout on read/list

When we request claim data, detect if it's the old layout and
automatically migrate all the claim data to the new layout. At the same
time detect older schema (e.g. when claim.Installation was claim.Name)
and handle that migration too.

When we cut to v1.0 we will stop supporting older claim data stores.

* Use a schema file at the root of home

Place a schema file at the root of home, schema.json, that says the
layout of the files used so we can detect if we should migrate.

{
  "claims": "CLAIM SCHEMA"
}

* Fill in the credentials spec version