Redo manifest collections as maps
Use custom types for manifest Outputs, Parameters and Credentials so
that they can be defined in the manifest as lists but worked with in the
code as maps. This way when we need to access an element by its name we
can do so without having to iterate over the entire array.