Migrate credentials
Migrate credentials during porter storage migrate:

* Track credential-set schema in schema.json under the "credentials",
e.g.   "credentials": "cnab-credentialsets-1.0.0-DRAFT-b6c701f"
* A migration will upgrade everything that is out of date, if claims
fail, it will keep migrating credentials then report all failures.
* Removed unused credential set migration from yaml to json. We haven't
used it for months and it isn't possible anymore with the filesystem
store / crud.Store interface.
* Renamed mybuns testdata to mybun for consistency and decency.
* Removed extra exclamation points in the migration logs. Now that we
are logging to a file, and using a dedicated command, grabbing people's
attention isn't as critical.
* Tweak when we log Migration Complete/Failed and the log file path so
that it is captured appropriately in the console and in the log file
without duplication.
* Aggregate errors from migration so that we continue on error but
report that an error occurred during migration.
* Implement interface methods in storage.Manager so that our custom
HandleConnect is called exactly when we want it, instead of the
HandleConnect of our datastore.