Support bundle dependencies with new claims spec (#1193)
* Generate parameter sources for bundle dependencies

* Inject dep outputs from parameter sources at runtime

* Update testdata

* unskip tests now that dependency output parameter sources is
implemented.
* update testdata to get tests passing

* Print the bundle when we can't load it

* Support setting params directly from dep outputs

parameters:
- name: PARAM
  source:
    dependency: DEPDENDENCY
    output: OUTPUT

* Resolve dep outputs at runtime

* Redo manifest collections as maps

Use custom types for manifest Outputs, Parameters and Credentials so
that they can be defined in the manifest as lists but worked with in the
code as maps. This way when we need to access an element by its name we
can do so without having to iterate over the entire array.

* Selectively build template data at runtime

Build template data at runtime to inject into mustach only for templated
outputs from parameter sources. Ignore parameter sources created for
parameters that have outputs directly wired into them.

* Incorporate review feedback

* Add function for building a prerequsites installation name for a
dependency.
* Use mysql bundle v0.1.3
* Handle injecting multiple outputs from a dependent bundle
* Use IsOutputSensitive helper method

* Update mysql bundle.json

* Update docs for wiring dependency outputs

* Remove padding from bottom of html lists

* Fix unit tests

* Set the cached bundle's manifest after pulling a bundle, not just when
it's a cache hit.
* Set the driver to debug for unit tests. We used to be able to rely on
it defaulting to debug in unit test mode, but now that we have multiple
test drivers, that doesn't work anymore.

* Update porter schema with source.dependency