Finally got it working
Now we have a hugo base image and we can pass `hugo` command when running the docker container  and the files are now watched so upon changes the  hot reload will  reflect the changes. Also moved the dockerfile to docs folder. Il think it make more sene to have there.
Also fixed the trailing that could cause issues with url redirection

Signed-off-by: Mohamed Chorfa <chorfa672@gmail.com>