WIP: Make docs preview run into docker (#1246)
* Make docs preview run into docker

Signed-off-by: Mohamed Chorfa <chorfa672@gmail.com>

* Update Makefile

Signed-off-by: Mohamed Chorfa <chorfa672@gmail.com>

* Apply code review

Signed-off-by: Mohamed Chorfa <chorfa672@gmail.com>

* Finally got it working

Now we have a hugo base image and we can pass `hugo` command when running the docker container  and the files are now watched so upon changes the  hot reload will  reflect the changes. Also moved the dockerfile to docs folder. Il think it make more sene to have there.
Also fixed the trailing that could cause issues with url redirection

Signed-off-by: Mohamed Chorfa <chorfa672@gmail.com>

* Apply code review comments

Signed-off-by: Mohamed Chorfa <chorfa672@gmail.com>

* update Preview documentation

Update `Preview documentation` section with the new way of rendering of the documentation via docker

Signed-off-by: Mohamed Chorfa <chorfa672@gmail.com>

* Update config.toml

Signed-off-by: Mohamed Chorfa <chorfa672@gmail.com>