Add GOVERNANCE file (#1284)
* Add GOVERNANCE file

This explains how the project governs itself today without attempting to
introduce any language that we didn't already have existing. Over time
we can iterate and improve upon it in follow-on pull requests. I want to
get this in without modifying our existing goverance for our inclusion
into the CNCF.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Incorporate Review Feedback

* Fix typo
* Link to sandbox

Co-authored-by: Vaughn Dice <vadice@microsoft.com>
Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Incorporate feedback

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

Co-authored-by: Vaughn Dice <vadice@microsoft.com>