Fix defaulting of reference on windows
We were accidentally using filepath when joining components of a docker
reference, which creates an invalid value registrybun (wrong slash).

The exiting unit tests would have caught this but we don't (yet) run
CI on windows. I just hit this bug when adding CI on Windows so this
will have coverage soon.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>