Revert images.contentDigest back to digest
This undoes an accidental change from
https://github.com/getporter/porter/pull/1308 that changed the manifest
field images.NAME.digest to images.NAME.contentDigest. This is a
breaking regression for anyone previously using images. There wasn't any
need to change the field name in the manifest, we only needed the
bundle.json to use the term contentDigest.

I have changed the manifest field back to digest, and ensured that the
user documentation for that field is up to date.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>