Use unique bundle names
Most of our test bundles have the same name, e.g. mybun,
but when we run in parallel that means we are overwriting the same
invocation images and the wrong image is used with the test.

Use a unique bundle name when copying a test bundle into the test
directory so that they can safely run in parallel.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>