Fix more tests on Windows
* Always make a path using GetHomeDir, or p.BundleDir, or p.TestDir to
account for root filesystem differences when running the tests on
Windows vs other operating systems.

e.g. The test home on Windows is C:\.porter and otherwise is /.porter

* Use CGO_ENABLED=0 when running tests

On windows there is a bug in go where gcc is required when running some tests. Not sure why but this
prevents go from trying to use CGO when running integration tests.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>