Use a custom extension for File type (#1420)
* Use a custom extension for File type

Previously we had checks throughout the codebase checking if a parameter
is Porter's special "file" type. Which lead to this bug.

I have made the file type a custom extension under the porter namespace,
and centralized the logic for the parameter type. I will submit this to
the CNAB spec as a well-known extension since I think it is useful
beyond Porter.

I also made our extension constants consistent.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Review feedback

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>