Publish mixins feed to /mixins/atom.xml
Originally we had only a single feed, just for mixins, and it was
located at /atom.xml. When we introduced plugins, we changed the
location of the feed to /mixins/atom.xml but for backwards compatibilty
reasons we continued to publish the file to /atom.xml and used a
redirect for /mixins/atom.xml.

That was over a year ago and at this point, no one should be using
clients that old. So this feels safe to move.

By publishing the file where it Porter looks for it when installing
mixins, we don't need a rule on our CDN to do that rewrite which makes
things less complicated for us.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>