Do not publish untagged or latest to mixin feed (#1466)
* Do not publish untagged or latest to mixin feed

The file structure in the bin is used when generating the mixin feed. So
if there is a folder named "latest" then a version "latest" is put in
the atom.xml. A previous commit caused an untagged release to be in the
bin, instead of being renamed to canary and uploaded to our feed which
broke installing porter.

I have moved the directory prep into mage, calling it in make xbuild-all
so that it is guaranteed to happen. I have also added a check in the
atom feed generation to ignore those releases if found.

Now after you cross build, your bin looks like this

bin/
  mixins/
    exec/
      canary/... OR
      v1.2.3/...

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix typos

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Make most recent build available from constant location

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Only generate for canary and tagged versions

* Flip the logic for the atom feed generation to ignore anything that
isn't canary or vX.Y.Z, this lets us put whatever we want in bin without
worry that it will end up in the feed.
* When we save the xbuild-all binaries, save them to bin/dev instead of
bin/latest to avoid confusion
* Found another packr directory that needed to be cleaned after build

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Include status text when mixin install fails

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>