Publish version binaries before install script
We just ran into a publish problem on the last build where the install
script (which hardcodes the latest version) was uploaded before the
binaries for that version. The upload for the binaries failed so the
install script was pointing at missing binaries.

I've changed the order so that we upload to /porter/VERSION/* first.
Then if that's successful, upload /porter/latest/install-*.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>