Do not use HTTP/2
We have been having intermittent problems with errors such as:
* 503 Service Unavailable
* HTTP_1_1_REQUIRED

Based on discussions in an Azure WebApps forum, people are suggesting that this is caused by Azure CDN's use of HTTP/2 which can be unstable with some clients.
Verizon's CDN has it enabled as well. I want to try turning it off to
see if that reduces or eliminates the errors.

I am disabling HTTP/2 on the client and in our install scripts

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>