Upgrade to gopkg.in/yaml.v3
I had a bug report for problems parsing a --yes flag as a bool instead
of a string, causing the flag to be passed as --true.

install:
- exec:
    description: "use a yes flag"
    command: az
    arguments:
    - something
    flags:
      yes: ""

This is fixed in v3 of the yaml package, becauase it uses a newer
version of yaml that doesn't interpret yes/no as boolean keys.

This version bump does change the indentation strategy of serializing to
yaml but other than impacting our test strings, it's still valid yaml,
just uses more spaces.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>