Add example for capturing output
I've realized that we need to start making our library a bit easier to
use programmatically. For example defaults set in the cmd/porter package
aren't used, so they need to start being defaulted in the Validate
command.

Ideally the command itself should do the validation and we should stop
doing PreValidate with cobra.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>