Bandaid on publish (#1484)
* Split publish into two steps

The build is failing on building the docker images because
of our CDN problems. This makes it easier to rerun just the docker
image part, without retriggering the Azure bug.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Build docker images using local binaries

Copy the local binaries from the bin instead of running porter's install
script. It will reduce build errors and is more accurate in case of the
CDN lagging.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Always build docker images after xbuild

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Add REGISTRY build variable

REGISTRY is where bundles and docker images should be pushed. It defaults to getporter for normal releases, and when we are doing a test release, it uses getporterci.

This lets us test ALL of a release without impacting any production assets.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Allow running a test release build for admins in a PR

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>