Capture yamux errors from plugin execution
When the plugins execute, there is a library (yamux) that is logging
directly to stderr things that don't appear to be errors. For example
when we disconnect the plugin, it sometimes logs "broken pipe".

2021/03/16 23:14:01 [ERR] yamux: Failed to write header: write unix @->/tmp/plugin968987118: write: broken pipe
2021/03/16 23:15:41 [ERR] yamux: Failed to write header: write unix @->/tmp/plugin806433500: use of closed network connection

These don't immediately appear to be real errors because the bundle
executes successfully and the data we expected to be persisted by the
plugin, is present. It's super hard to reproduce so I am not not 100%
confident that there may not be an underlying bug.

I have patched hashicorp/go-plugin to redirect yamux logs to the plugin
client's logger. This lets us capture the error and look at it further
if we need to later.

The log emitted by yamux is plaintext and I can see these logs when I
use the logging Exclude function. However I never see them in
logPluginMessages where I am handling the logs. I'm not sure why but
since we wanted to NOT print them out anyway, it seems okay for now.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>