Allow overriding home dir and download url during install
When installing, let the user override with PORTER_HOME and
PORTER_MIRROR the location where porter should be installed, or the
location from which it should be downloaded.

I have updated the porter CLI to understand the new configuration value
PORTER_MIRROR, which and also be specified as a flag or config file
setting.

This lets us test out alternate download infra from end-to-end, and also
allows a user to set up a mirror of Porter's assets.

I am also getting rid of the troubleshooting I had added for Azure's
storage and CDN now that we aren't using it anymore (e.g. dynamically
generated PORTER_TRACE user agents).

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>