Do not copy extra junk when installing porter
The install target is a bit lazy and copies everything in certain
directories. This results in files that do not need to be installed
copied into PORTER_HOME. For example, the release directories (dev,
latest, tagged) for the exec mixin.

The extra binaries are a problem because it seriously bloats invocation
images built on a dev machine.

This moves the logic to our magefile, and is more explicit about what we
want to copy.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>