Only lint example bundles for pull requests
To help speed up PR turnaround, just lint example bundles and make sure
that they have a correct porter manifest. Nothing about the example
bundles really exercise edge cases, they aren't test and we just want to
make sure that when we change Porter, the examples don't get out-of-date
which linting will catch.

In a follow-up I will have the build conditionally build the bundle if
it was modified.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>