Add updatetestfiles target
Some of the tests, notoriously the schema test, relies on a "golden"
test file. The test output is compared against that file to determine if
the command output of `porter schema` is correct.

Changing the schema is difficult because you have to know how to update
that test file.

I have updated the test to detect if the schema has changed and prompt
you to run this new target `mage updateTestfiles` if you intended to
change the schema. It will handle updating the golden file to match the
new output of the test for you.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>