Fix using kind in integration tests (#1575)
* Fix integration tests

This moves the cluster logic logic into a reusable mage package. People
can override some values, like the name of the kind cluster, with
KIND_NAME, etc.

I've also removed the path filters for the integraiton tests. They
cannot be both triggered by a path, and triggered with a GH comment. We
should just run it manually when it is appropriate.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Clarify when the integration tests run

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>