Load defaults before applying config to flags
I've added a regression test to ensure that when we set a flag for
something that falls back to the config, but it also has a flag default
set, that the config value is properly bound to the flag.

Fixes the edge case of bad config data and the user running porter build
without specifying the --driver flag. Previously it was defaulting to
the flag's default instead of binding to the config value.

I've also tweaked the TestPorter so that it doesn't accidentally load
the config and env vars from your dev machine when running tests.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>