Add new build driver for buildkit (#1567)
* rename build provider
* Add new build driver for buildkit
    * Add experimental flags to config file. These features are disabled
by default.
    * Add experimental flag for "build-drivers". This allows specifying the
build driver, e.g. docker, buildkit, in the config file.
    * Add driver for buildkit.
* Add --experimental flag and refactor config/data
    * Add --experimental flag that accepts a comma separated list of feature
flags. Right now we only support "build-drivers".
    * Refactor porter's config and data structures so that we can't get nil
pointer exceptions or try to access the configuration data before it is
loaded.
    * Move helper functions to Config so that they are easier to find.
    * Remove helper functions for config values that aren't needed anymore
because its safe to directly access the bound field going forward.
    * Use a fork of viper with a patch for a bug in how we load
configuration when a config file is not found. We should still bind to
environment variables.

* Fix review feedback
    * Fix setting defaults on viper config
    * Add missing example to custom dockerfile page
    * Fix comments
    * Bind --driver to the command flag, not the config struct directly
    * Add check to AddTestDirectoryFromRoot

* Load defaults before applying config to flags

I've added a regression test to ensure that when we set a flag for
something that falls back to the config, but it also has a flag default
set, that the config value is properly bound to the flag.

Fixes the edge case of bad config data and the user running porter build
without specifying the --driver flag. Previously it was defaulting to
the flag's default instead of binding to the config value.

I've also tweaked the TestPorter so that it doesn't accidentally load
the config and env vars from your dev machine when running tests.

* Support setting experimental flags programmatically

Make it easier for tests or downstream consumers of our library to
enable experimental flags quickly without having to understand our
configuration system.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>