Do not resolve prerelease versions with latest
Now that we will have prerelease versions of the exec mixin available,
porter mixin install exec is failing, because it's resolving the mixin @
v1-canary. Latest should resolve to the most recent _stable_ version
available, which excludes the -canary entries.

I also fixed the sorting of the entries in atom.xml so that when two
binaries have the same timestamp, they sort consistently by name and
version.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>