Use the specific version in install scripts
We used to take a permalink for the mixins in the installation script,
because each had it's own release cadence. Now that we only install
porte and exec, we can use just the version in the install script.

This is important because we don't publish canary/latest entries into
the mixin feed for pre-releases.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>