Fix applyTo check for outputs in porter explain
I have consolidated the logic for checking if a scoped item: parameter,
credential or output, should be included in the output of explain to
avoid errors when duplicating the logic for each one.

There was also a copy/paste of the appliesTo logic for
ParameterDefinition (our def of a parameter in the porter manifest) that
I updated to implement Scoped so that we can hopefully avoid more
problems with this type of check.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>