Fix validation when no bundle specified
We have a sentinel error in our code that checks if we made it into
executing a bundle, when no bundle is specified. I managed to trigger
that today on the v1 branch with the following command:

porter install -c mycreds

I was not in a bundle directory. I've added a check for when the user
doesn't specify anything so that we can provide a better error message
and stop early.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>