Fix check for if a bundle was provided
Install requires a bundle definition from either --reference, --file,
--cnab-file or from the bundle in the current directory. The other
commands can either use those or use the installation name to look up
the bundle definition.

I found a way to get install to execute without specifying a bundle
definition and it blew up later in code that execpted that this had
already been validated. The command was

porter install --creds aks

I didn't specify a --reference or anything and wasn't in a bundle
directory.

Hopefully this plugs the gaps in wher we check for this condition.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>