Add ExtendedBundle wrapper
This wraps the cnab-go bundle and adds support for custom extensions
that porter supports. It will help us wean off of using the porter
manifest at runtime, which is a huge problem because then our support
for spec extensions only works when it was authored in Porter.

More work is needed to fix this, but this is just a refactoring to get
the data in the right place, i.e. when we do pass around a bundle, it's
a bundle that knows how to deal with extensions.

I am still saving a cnab bundle document so that if we change the
structure of the wrapper it doesn't impact the stored document.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>