Always pull the invocation image
Sometimes the invocation image only exists in the local docker cache but
was never pulled. In that case, there isn't a repository digest for the
image in the cache and it will fail validation.

Always pulling, will ensure that the repo digest is always populated to
avoid unexpected errors about missing digests.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>