Allow repeating install when --force is specified
When building out a bundle for the first time, or troubleshooting,
sometimes it is incredibly helpful to be able to repeat the install
action, even when it was successful the first time.

For example, I am iterating on the install action for a bundle, adding
new steps each time I get the bundle to install successfully. In this
case, it makes sense to allow the bundle author to repeat the install
action with --force.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>