Always collect outputs even when bundle fails
When a bundle fails, we still need to collect any generated outputs
because they may contain state that is needed for the next run.

What I am running into is that when I install a bundle that uses
terraform, and it fails, the tfstate file output is not persisted, so I
can't repeat the install command again.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>